+++ src/balanced/red_black_node.h
@@ -5,7 +5,7 @@ enum Color { RED, BLACK };
 
 template <typename T>
 struct rbnode {
-    rbnode(T data, int key) {
+    rbnode(T data, long key) {
         this->data = data;
         this->key = key;
         left = right = parent = nullptr;
@@ -13,7 +13,7 @@ struct rbnode {
     }
 
     T data;
-    int key;
+    long key;
     bool color;
     rbnode *left, *right, *parent;
 };
